- Rendu la destruction de base de donnée lors des tests unitaires plus robuste.
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2010/Epsitec.Cresus@15173 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Cresus.Database.UnitTests/Helpers/DbInfrastructureHelper.cs
@@ -19,22 +19,19 @@
 
 		public static void DeleteTestDatabase()
 		{
-			// TODO Have a better method to drop the database, that does not require to connect to
-			// it before, and that does not fail if the database does not exists or is invalid.
+			DbAccess access = TestHelper.GetDbAccessForTestDatabase ();
 
-			using (DbInfrastructure infrastructure = new DbInfrastructure ())
-			{
-				DbAccess access = TestHelper.GetDbAccessForTestDatabase ();
-
-				infrastructure.AttachToDatabase (access);
-				infrastructure.DropDatabase ();
-			}
+			DbInfrastructure.DropDatabase (access);
 		}
 
 
 		public static void ResetTestDatabase()
 		{
-			DbInfrastructureHelper.DeleteTestDatabase ();
+			if (TestHelper.CheckDatabaseExistence ())
+			{
+				DbInfrastructureHelper.DeleteTestDatabase ();
+			}
+
 			DbInfrastructureHelper.CreateTestDatabase ();
 		}
